ignore incomplete HTTP transactions
If only the response gets captured without the request, this
resulted in a NullPointerException. Since such responses
naturally tended to be at the beginning of the capture, the
rest of the (complete) HTTP transactions were "lost". After
this commit, such incomplete entries are ignored, so at
least the rest is read without problems.